Don't allow OOBS_SERVICE_IGNORE status for an init script
This status was handled internally as "remove from list", which in turn made the backends remove all symlinks in /etc/rcN.d/ for that script and runlevel. This is not the recommended way since upgrade of the package will re-enable the script automatically. Clients must use OOBS_SERVICE_STOP instead.