Use service priority value 0 to keep previous value
Most of the time, we want to enable/disable a service, but keep its previous priority, instead of changing it to an arbitrary value. This will be used by services-admin, which will always pass a null value. If the script was not present in this runlevel, the null priority will be sent to the backends, which will set it to the default for us (50 ATM).