Increase OOBS_[UG]ID_MAX to G_MAXINT32 when possible and install oobs-defines.h
The system-tools-backends use an int32 to store [UG]IDs, so we need to respect these limits even if we and the system support it. But using a hard-coded limit of 65,535 makes no sense. If uid_t and gid_t are big enough, we need to allow these values, else we could destroy the system's configuration. Use a custom computing method to get the max value [ug]id_t can hold, since no system constant exists for that.