Add oobs_object_authenticate() and GError reporting
The System Tools Backends 2.9.1 support a new authenticate() method called for all modules, with the interface org.freedeskop.SystemToolsBackends.Authentication. Support this via a new oobs_object_authenticate() function that works generically for all OobsObjects. Remove oobs_object_get_authentication_action() which was only used with OobsSelfConfig, since we should not have to deal with PolicyKit directly.

Add general support for GError reporting, which is only used in the new method for now. Update documentation to show it.