Fix SelfConfig, add convenience function
We need to get the UID on start to find the right user, and to send it to the backends. Else we were sending a wrong UID to the backends, leading to silent failure.

New oobs_self_config_is_user_self() convenience function that simply compares pointers, it's cleaner in the clients.