Don't remove unknown users from groups
Users not listed in /etc/passwd can still be valid ones (e.g. LDAP). The backends send us a raw list of usernames, we now keep it as a basis for commits, and only add/remove to it known users. OobsGroups now keep two users list: one with usernames, and one with references to OobsUsers, which is used for the public API. Unknown users are kept in the first list and get eventually committed without being exposed anywhere.

This requires us to remove manually users from all groups when removing them, else we would commit their names to /etc/group depite them not existing anymore. This is done in oobs_users_config_delete_user().

https://bugzilla.gnome.org/show_bug.cgi?id=608815