Ensure passwords are erased from memory
We need to ensure clear text passwords for users and groups don't stay forever in memory. This means we erase them as soon as objects are committed, and on finalize() in case objects have not been committed. One drawback could be that a failed commit can't be reproduced without re-setting the password field, but no clients should rely on that behavior.