Fix crash when updating with old OobsObjects destroyed
Destroyed OobsObjects didn't remove the D-Bus filter function, which was called when *any* object was updated. Since the reference was now invalid, we got a crash. This is likely to have een revealed by recent changes that allow OobsObjects to be actually destroyed after an update has made them obsolete.