Clear runlevels list before updating
Before that, we were adding the 6 runlevels again and again, on each update. This lead to bugs when committing services changes, because only the first occurrence was modified, and others were always set to 'stop'. Since each runlevel was committed several times, the last one was deciding of the final state of the service, which was 'stop'.

The only known symptom of this was in time-admin, which is updating the services list when installing NTP support. See
https://bugs.launchpad.net/ubuntu/+source/gnome-system-tools/+bug/551936