Set D-Bus timeout to max for async replies
Some tasks can take a long time to complete, and the backends
don't reply until they are done. Increase the timeout when
committing asynchronously to avoid getting an error. This means
we can end up with an infinite progress bar and no error, but
that's as helpful as a timeout error, and it's the only solution
without a deep refactoring of the backends.