Store OobsUser's main group as GID rather than OobsGroup
Storing main group as a reference to an OobsGroup means we have to
keep track of OobsGroupsConfig updates, and mess with reference
cycles. The most accurate information provided by the system is
the GID of the main group, so it's better to use this and retrieve
the corresponding OobsGroup when oobs_user_get_main_group() is called.

This fixes issues with OobsGroupsConfig not being updated properly when
creating a user triggers the creation of its main group in the backends:
we couldn't update OobsUsersConfig and OobsGroupsConfig correctly since
they each depended on the other, which means the new main group was not
found, and made subsequent commits fail (invalid GID). See e.g.
https://bugs.launchpad.net/ubuntu/+source/gnome-system-tools/+bug/659758